refactor: entity 패키지명을 domain으로 변경
