+++ src/main/java/roomescape/domain/reservation/ReservationSearchFilter.java
@@ -1,9 +1,8 @@
-package roomescape.dto.reservation;
+package roomescape.domain.reservation;
 
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Objects;
-import roomescape.entity.reservation.Reservation;
 
 public record ReservationSearchFilter(Long themeId, Long memberId, LocalDate dateFrom, LocalDate dateTo) {
 
@@ -12,6 +11,9 @@ public record ReservationSearchFilter(Long themeId, Long memberId, LocalDate dat
     }
 
     public List<Reservation> doFilter(List<Reservation> reservations) {
+        if (isNeeded()) {
+            throw new IllegalArgumentException("[ERROR] 필터링이 불가능한 요청입니다.");
+        }
         return reservations.stream()
                 .filter(reservation -> Objects.equals(reservation.getTimeId(), themeId))
                 .filter(reservation -> Objects.equals(reservation.getThemeId(), themeId))
