+++ src/main/java/roomescape/domain/reservation/ReservationTime.java
@@ -1,4 +1,4 @@
-package roomescape.entity.reservation;
+package roomescape.domain.reservation;
 
 import java.time.LocalTime;
