+++ src/main/java/roomescape/domain/reservation/Theme.java
@@ -1,4 +1,4 @@
-package roomescape.entity.reservation;
+package roomescape.domain.reservation;
 
 public class Theme {
