+++ src/main/java/roomescape/dto/reservation/ReservationTimeRequest.java
@@ -2,7 +2,7 @@ package roomescape.dto.reservation;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalTime;
-import roomescape.entity.reservation.ReservationTime;
+import roomescape.domain.reservation.ReservationTime;
 
 public record ReservationTimeRequest(
         @JsonFormat(pattern = "HH:mm")
