+++ src/main/java/roomescape/infrastructure/member/MemberInfo.java
@@ -1,6 +1,6 @@
 package roomescape.infrastructure.member;
 
-import roomescape.entity.member.Role;
+import roomescape.domain.member.Role;
 
 public record MemberInfo(Long id, String name, String email, Role role) {
