+++ src/main/java/roomescape/service/memeber/UserReservationTimeService.java
@@ -4,8 +4,8 @@ import java.time.LocalDate;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import roomescape.domain.reservation.ReservationTime;
 import roomescape.dto.reservation.ReservationAvailableTimeResponse;
-import roomescape.entity.reservation.ReservationTime;
 import roomescape.repository.reservation.ReservationRepository;
 import roomescape.repository.reservation.ReservationTimeRepository;
