+++ src/main/java/roomescape/service/reservation/ReservationService.java
@@ -3,14 +3,14 @@ package roomescape.service.reservation;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import roomescape.domain.member.Member;
+import roomescape.domain.reservation.Reservation;
+import roomescape.domain.reservation.ReservationSearchFilter;
+import roomescape.domain.reservation.ReservationTime;
+import roomescape.domain.reservation.Theme;
 import roomescape.dto.member.UserReservationRequest;
 import roomescape.dto.reservation.ReservationRequest;
 import roomescape.dto.reservation.ReservationResponse;
-import roomescape.dto.reservation.ReservationSearchFilter;
-import roomescape.entity.member.Member;
-import roomescape.entity.reservation.Reservation;
-import roomescape.entity.reservation.ReservationTime;
-import roomescape.entity.reservation.Theme;
 import roomescape.exceptions.reservation.ReservationDuplicateException;
 import roomescape.infrastructure.member.MemberInfo;
 import roomescape.repository.member.MemberRepository;
