+++ src/test/java/roomescape/service/reservation/ReservationServiceTest.java
@@ -11,11 +11,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.jdbc.Sql;
 import org.springframework.transaction.annotation.Transactional;
+import roomescape.domain.member.Role;
+import roomescape.domain.reservation.ReservationSearchFilter;
 import roomescape.dto.member.UserReservationRequest;
 import roomescape.dto.reservation.ReservationRequest;
 import roomescape.dto.reservation.ReservationResponse;
-import roomescape.dto.reservation.ReservationSearchFilter;
-import roomescape.entity.member.Role;
 import roomescape.exceptions.EntityNotFoundException;
 import roomescape.exceptions.reservation.ReservationDuplicateException;
 import roomescape.infrastructure.member.MemberInfo;
